+++ client/src/components/Analytics/Analytics/index.ts
@@ -1,0 +1,3 @@
+import QueryData from './QueryData';
+
+export default QueryData;
