+++ client/src/components/Analytics/Query/index.ts
@@ -1,0 +1,3 @@
+import Overview from './Overview';
+
+export default Overview;
