+++ client/src/components/Analytics/index.ts
@@ -1,0 +1,3 @@
+import AnalyticsRouting from './AnalyticsRouting';
+
+export default AnalyticsRouting;
