--- client/src/components/Analytics/index.tsx
@@ -1,3 +1,0 @@
-import Analytics from './Analytics';
-
-export default Analytics;
