+++ client/src/components/Auth/GetAuth.tsx
@@ -4,7 +4,9 @@ import Login from './Login';
 import Register from './Register';
 
 const GetAuth: FC = () => (
-  <div>
+  <div className="slim">
+
+    <img className="loginLogo" src="../../image/Group3.png" />
     <Login />
     <Register />
   </div>
