+++ client/src/components/Home/Home.tsx
@@ -5,7 +5,7 @@ import PlaygroundDisplay from './PlaygroundDisplay';
 
 const Home: FC = () => (
   <>
-    <Link to="/analytics">Analytics</Link>
+    <Link to="/analytics"><button type="button">Analytics</button></Link>
     <PlaygroundDisplay />
   </>
 );
